LayoutTests/fast/dom/getElementById-with-rope-string-arg.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("This tests that calling getElementById() with a concatenated string as the argument doesn't make us crash on subsequent access to that string.");

var foo = "foo";
var foobar = foo + "bar";
var nonexistentID = "something" + foobar;

document.getElementById(nonexistentID);

// Test passes if this line does not cause a crash:
shouldBe("nonexistentID", "'somethingfoobar'");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
