Safari-609.1.11.1/LayoutTests/fast/dom/getElementById-with-rope-string-arg-expected.txt

This tests that calling getElementById() with a concatenated string as the argument doesn't make us crash on subsequent access to that string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nonexistentID is 'somethingfoobar'
PASS successfullyParsed is true

TEST COMPLETE
